Restructured the GEM
