+++ lib/gitget/github_developer.rb
